Cross-compiled build with D-Bus dependencies.
We use GDBus from the GNOME project. GDBus is part of GIO and
therefore we need to link against a bunch of libraries. This commit
works for local Linux PC and cross-compilation for Raspbian.

For Raspbian, the PKG_CONFIG variable needs to point at some
pkg-config binary v0.26 or higher. This is required because the
current Linaro toolchain from https://github.com/raspberrypi/tools.git
ships with pkg-config 0.25, but this version does not understand the
PKG_CONFIG_SYSROOT_DIR environment variable. To avoid hacks, we simply
require v0.26 or higher.