Sending playback start and stop over D-Bus works.
Upon reception of DRCP command codes 0xb2 or 0xb3, a playback stop or
playback start signal, respectively, is sent over D-Bus to anyone who
is listening.

To make this work, the D-Bus I/O is done in a GThread. Alternatively,
we could have iterated the GMainLoop in our own main loop, but this
seems to be too much of a disturbance of the loop. It would have ended
with pulling in lots of GLib-specific stuff to get the blocking and
timeout handling right, and it simply isn't worth it.