Simplified transaction memory management.
The information whether or not to free a transaction object after use
is now associated with the transaction object itself. Instead of using
a single flag whether or not to free the current active object, dcpd
now relies on information provided by the objects. This is simpler and
more robust at the cost of using a bit more RAM.