Add function for sending registers to slave.
This works by simulating a slave read request. We fill in the
transaction data using the register definition (just like we are
generating master transactions for DRCP) and skip the transaction
state that would read this data from SPI. The synthesized request can
then be processed as any other transaction.