Add options for setting MAC and name of primary network interface.
It is actually pretty hard to find out which exact network interface
_is_ actually the primary network interface. We cannot just use "eth0"
because of udev and undefined enumeration of devices by the kernel. It
is well possible for eth0 to be some external interface attached to a
USB if it was enumerated before the one that we would expect.

To make a long story short, finding the "primary network interface" is
a matter of definition and also a very system-specific problem. We
leave this work to others (probably some system-specific shell script
wrapper) and simply rely on information passed to us on the command
line.

Change-Id: I5df4f760176f634d0cbae2155f48d9af4616dd81