Rename files to avoid future confusion.
There are two areas concerning networking right now:

    1. Network configuration (set/get IP address etc.); and

    2. DCP over TCP/IP (as an alternative command interface that
       replaces the SPI interface).

A third area is about to be implemented next: The TCP/IP tunnel for
the T+A device control smart phone app.

The source file names are more specific now so that it is easier to
tell which file belongs to which facet of networking.

Change-Id: I7c439651c478e5088df414c2f6adb9e31112f7b8