Incomplete DCP integration of WLAN site survey (#133).
Implementation of registers 104 and 105 has been added, based on the
new WLAN site survey code. A survey can be initiated and valid XML
data is generated from the results.

The remaining problem now is the size of the XML data. We must scatter
that data over several SPI transactions, much like we are doing for
DRCP traffic already (which is handled in a special way). There is
currently no code that does this for regular DCP packets outside of
the DRCP domain. The basic building blocks are already available,
however, and used by the DRCP code; now it is just a matter of finding
a good point where to handle DCP this way as well.

Change-Id: If94242750afbb5a7b9a10602cb1f8d033ea804e9