Fix bug in code that calls DCP read handlers.
Signed vs unsigned bug that caused -1 returned from a failing read
handler to be interpreted as SIZE_MAX, which in turn got truncated to
UINT16_MAX, which was used as the size of the payload to be sent to
the SPI slave. The data sent was mostly random junk from the heap,
nicely packaged into many DCP fragments. Of course, random crashes due
to page violations could have been possible as well.

Bug was introduced in 7152174 as a result of a required refactoring
and not checking the Cppcheck report...

Change-Id: Ia82221eba2c186a69de2c2780ee64c461c650b6e