Enable triggering updates, the hacky way (#108).
Updates are triggered from a generated shell script. The script runs
opkg (using sudo) and reboots if the upgrade succeeded (using D-Bus,
regardless of the number of packages updated).

We are using an external script that does not belong to any package so
that opkg cannot kill us as part of a package deinstallation routine.
This way, upgrading dcpd will work as well.

Change-Id: I440bd54df12101d9a1f1e0e64701046cee196a3f