Wake up WLAN hardware at certain strategic points.
The hardware is woken up once during program startup, and every time
the list of services (as reported by Connman) changes. The latter
makes sure that we enable WLAN when (1) the Ethernet cable is removed
or (2) we are losing WLAN networks, possibly as a result of WLAN
hardware powering down. This basically keeps the WLAN hardware powered
all the time.

Change-Id: Ife85dd49574d22737be2e031de08a20c9d852cf5