Fix requesting Airable authentication URL.
The parameters were swapped. Unfortunately, there is a mismatch
between the order of parameters in the D-Bus interface on the one hand
and in the app protocol on the other hand. Not very beautiful, but
keeping the interface the way they are and just fixing the only caller
in the whole code base is the best solution.

Change-Id: Iba7ca7d33acc8c70765139a5874105600f942b40