Parse meta data written to register 78.
We use ASCII control character GS (0x1d) to separate title, artist,
and album name, in that order. Use of GS is optional.

It is perfectly fine not to use GS, thus to send flat strings. In this
case, the old behavior of using the string as plain stream title is
retained.

Change-Id: Ifdecf48b7adf4c2dde3b0d7244e1289d8c009a1b