ConnMan agent can look up passphrases for queries.
We use a new kind of configuration file for this. It is a
configuration file of our own, not seen by ConnMan. Requests are
fulfilled from that configuration file, and data entered by the user
should be written to that file (not implemented this way yet). ConnMan
uses the information read from our own configuration to fill and
manage its own set of configuration files.

This commit adds support for reading our configuration file, but it is
not created or written to yet (implementation was tested by editing
the file on a running system).

Any code for generating configuration files directly for ConnMan or
reading such files is going to be removed from dcpd soon. We are not
going to mess around with ConnMan configuration files anymore.

Change-Id: Ic08bd1959956ba725eebd533b1ad4ef70e9b33f2