Connect to configured WLAN when necessary.
This is, unfortunately, a rather big commit. It changes quite a few
things, and adds a lot of code for monitoring ConnMan service changes
and reacting to these changes.

Still, the commit does not fully implement network configuration.
There are two worlds: The old world which attempts to generate ConnMan
configuration files directly, and the new world which attempts to
configure ConnMan services via D-Bus.

Both worlds are running in parallel in this version, and together they
create a big mess.

Change-Id: I903346c264187c53e54a1b8253d3ae5ff04838d7