Port register code to new network preferences code.
This is a straight port to the new code for writing network
preferences. Unit tests have been ported as well, and they all pass.

Still, this version does not work properly because changes to the
network preferences does not trigger network reconfiguration. That is,
no notification is ever sent to ConnMan, meaning that the new
preferences may only take effect when ConnMan informs us about service
changes, not sooner.

Change-Id: I552f70eb7b9834580d29a77dc071ebc26a302f5b