System update script can switch between versions.
We need to able to switch distributions between versions so that at
least we can switch back and forth between stable and testing. This
involves enforced upgrading, installing, and removing packages as
defined by the distribution.

Since this use case seems to be not properly supported (see
https://groups.google.com/forum/#!topic/opkg-devel/C4SAIfFWp9k), there
are a few problems our update script needs to solve:

1) If the target distribution adds a root package (a package that no
    other package depends on), then the standard package dependencies
    are not enough to let opkg know about this. We need to install the
    added packages "by hand". For dependent packages, the dependency
    graph will take care of this.

2) Similarly, if an installed root package does not exist in the
    target distribution, then we need to remove it by hand in case.

3) opkg's --force-downgrade option does not seem to work with opkg
    upgrade, at least not in the intuitive way.

To solve the first two issues, the script generates a list of
installed packages along their version numbers and fetches the list of
packages defined by the target version (manifest file for the rootfs
image as generated by Yocto Project). These two lists are matched to
generate three lists of packages: those to be updated, those to be
installed, and those to be removed.

To work around issue number 3, the set of packages to be updated is
manually removed from opkg's status file. This way opkg fails to
figure whether or not the installed version is newer than the version
of the target distribution, and always reinstalls all changed packages
(note that all opkg's promising force options do, in fact, not work
for this). This requires root privileges, thus extension of sudoers
configuration.

Thank god we have a working recovery system (you're welcome...).

Change-Id: I60e7910bb189e9145f1435287927090215432f8e