Do not try to set DNS configuration if DHCP is on.
Actually, we are still trying to override DNS settings, but only in
case there are any nameservers defined in our configuration file.
(This feature is supposed to fix bad setups or temporarily enter a
different DNS in case the usual one in unreachable. The user
interface, however, does not support this.) This commit makes sure
that we leave alone DNS settings assigned by DHCP if the configuration
file does not define any nameservers.

Change-Id: I4bbd2a6d338edf48bb502341912067cb09684952