Explicitly disable IPv6 in network service configuration.
By default, all services are set to automatic configuration, meaning
DHCP for IPv4 and IPv6. In previous versions, we only explicitly
configured IPv4, but did nothing about IPv6. In a networking
environment that provides DHCPv6, this behavior results in a fully
working IPv6 connection.

This is actually amazing and should not be a problem, but ConnMan
reports "online" state as soon as any service goes online with any
protocol, and the Airable list broker uses that signal to start
loading/reloading the root list. It therefore tries to use whatever
connection is available to connect with Airable, but frequently this
doesn't work. It would be possible to fix this within the list broker
so that it explicitly checks IPv4 state, but such kind of
specialization to IPv4 would pointlessly work against IPv6 and
complicate things in case we add IPv6 support later.

So, since our appliances do not provide any IPv6 configuration
options, the only thing we can do for now is to always switch off
IPv6. This is what this commit does.

Change-Id: I72968036becce95319964913a4a36f7ecbaf7009