Fix more connectivity problems in conjunction with ConnMan.
ConnMan is a BITCH on D-Bus. Really. It would be hard to find software
as petulant and hard to interact with as this thing. The API looks
nice and clean in documentation, but the implementation has many
quirks that are NOT found in documentation.

ConnMan won't accept setting the auto-connect flag for services not
marked as favorite. The only way to mark a service as favorite is to
manually connect to it. Removing the favorite flag is not possible.
Setting properties such as auto-connect or IPv4/IPv6 methods to the
same value or trying to disconnect from an already disconnected
service is not only considered an error (which by itself is very picky
already, to say the least), no, ConnMan also errors out so hard so
that certain automatisms simply stop working! WTF?!

ConnMan is great if left alone, but interaction with it is
unnecessarily complicated, painful, and generally error-prone. These
are mere software bugs, of course, and they should be fixed within
ConnMan. This, however, is nothing we could do right now.

Change-Id: I1b723fc2d5ef912de7f4ab4227a9625c8739af14