Attempt to fix up system update state if necessary.
The system could be in a broken state because of a power failure or
some network failure during update, packaging errors, bugs in opkg,
etc. The list of possible failures is very, very long.

We are trying to detect a bogus system state (from the point of view
of the package manager) and continue/restart the update process as
seen necessary.

Note that this commit adds a feature that should be implemented in a
dedicated update program (such as the planned UpdaTA program). Once
such an update program gets deployed, it will be possible to avoid the
automatic check for bogus system states by creating the empty file
/tmp/dcpd_avoid_update_check.stamp before starting dcpd.

Also note that the generated update script now checks the content of
environment variable FORCE_SYSTEM_UPGRADE and the existence of empty
file /tmp/force_system_upgrade.stamp to enforce an upgrade being
performed. To enforce an upgrade by environment variable, set
FORCE_SYSTEM_UPGRADE to "force".

Change-Id: If021df63f2bafd2d6ba49a977575ec4da6252776