Allow overriding package repository configuration.
This feature is meant for testing purposes. Having to redirect web
traffic to a test server via DNS and/or having to take some hardcoded
distribution name from the SPI slave is simply very annoying. Testing
is made unnecessarily hard and unenjoyable this way.

This commit allows placing an override INI file next to the real one.
The real one gets read and updated as usual, but the strings written
to the generated package feed configuration files are taken from the
override file (which is never written to). Note that there exists no
code that would create such an override file. If it exists, then it
was created manually and on purpose by a developer.

Change-Id: I03cf7a4ea5ad1df3e6aebf370bda7d70e2699be2