Work around a bug in ConnMan triggered on first start (#471).
The bug can be triggered by booting a fresh system so that ConnMan
does not have any own configuration files yet. ConnMan will try to
connect with IPv4 and IPv6, but we are telling it to not use IPv6. So
it happens, but the global states drops to idle.

The network configuration is actually still OK, it's only the state
that is wrong. It only happens when the system is booted for the first
time, so we are working around this by disconnecting and reconnecting
to the LAN after first boot.

Let's hope that we can remove this workaround soon...

Change-Id: I2d3567abd8ae573cddcdbb4fa05d4ee7a805b215