Configuration values can be adjusted when read/written.
This change affects the convenience functions which operate on values
represented as string. Between the conversion from/to string and
writing to/reading from configuration, a callback function may be
called which adjusts the value as required.

Our use case is conversion between kilobits per second and bits per
second in the implementation of register 95.

Change-Id: I4ef457c5727b4c826279d776282cad22419b864f