Revert DRCP view switching behavior (see 90289c3).
Partial revert of 90289c33b604d0014b614aae82bb88e545217d20.

The problem with associating mere view switching with audio source
selection is that stopping playback would be enforced in various
situations, in obvious ones as well as in some non-obvious ones.

Even if this would be something we wanted, then that behaviour should
not be implemented in this case. There would be no way for the UI to
distinguish between simple temporary view switches and true audio
source selection. These concepts are kept separated again.

Change-Id: Ic53f6f616189c794573f1180f2f8d41c3962c34d