Override Ethernet IPv4LL in case WPS was requested.
Sometimes, Ethernet is auto-configured and won't move away. To address
this case, it is possible to just request WPS (was possible before,
but got ignored with a log message) and switch over to WLAN mode,
overriding the current Ethernet configuration.

The override only takes place if IPv4LL configuration is discovered,
otherwise the WPS request will be ignored as it has been before.

Change-Id: I5d09c39e841d48902899364192fd1cb6387161ca