WLAN site survey on Ethernet loss if necessary.
Before, we simply tried to enable power, but this turned out to be not
enough. Thing is, there might be no networks despite the power being
enabled (according to ConnMan), probably because ConnMan simply won't
see (or tell us) about WLAN suspend modes. A scan in this situation,
however, does the trick.

Change-Id: Ia6ee5bde0549496ca9561168a4336901330fc07b