Avoid GLib warning at startup.
The configuration may be updated very early during startup, even
before the D-Bus stuff is up and running. At this time, the
configuration changes are to be emitted to the system as a D-Bus
signal, but since there is no D-Bus interface yet, signal emission
fails with a GLib warning.

Successful emission of the signal at this point of startup is not
critical because processes interested in the values should read them
out explicitly anyway. Thus, we simply avoid the signal emission while
impossible.

Change-Id: Idb5321bcd18f10b312765e39b94efd6978621730