Add device ID to daemon configuration.
The device ID is either the Ethernet MAC where available, the WLAN MAC
on CalaBerbel, or empty in case of undefined appliance or if there is
no MAC address. The Airable list broker, for instance, may make use of
this.

Note that the device ID is not restricted to be a MAC address. In
fact, it may be any string, so we could as well put appliance serial
numbers read out from somewhere else or randomly generated IDs.

Change-Id: I3646cc98986b7c9c4e01c8b55bce137ec27c44bd