Add appliance-specific fallback networking technology.
In case there are no active WLANs (suspend mode or because there are
really none) and there is no Ethernet connected, we cannot determine
the correct technology to use for configuration. We should pick a
default technology in this case, but the correct default is specific
to appliances.

Therefore, we allow setting an appliance-specific fallback technology
and use it where necessary.

Change-Id: Ibc4a0c9632aa47e630a6e1c65e1202ddd004d277