Rename app audio source ID to "strbo.plainurl".
The old name, "App", did not look like the other audio source IDs,
plus the interface feeding this audio source is not necessarily an
app. In fact, all we known is a plain URL that shall be played, and
anyone can tell us one; hence the new, more generic name.

Change-Id: Ibf164817eba5ccd47f258692f1a98bea67c978f3