Support for receiving long, fragmented messages (#666).
In case a DCP packet of size 256 is received (the maximum possible
size), it is considered a fragment of a long message. The DCP
transaction logic transparently concatenates such long messages until
a packet with a size smaller than 256 is received (size 0 is a valid
option as well). Only then the payload is forwarded to the
corresponding register handler.

Change-Id: Ic80aac9c260041654f1c3af236b49233e3f1be1b