Increase URL buffer size to 2048, fix register sizes (#348).
The URL buffer size was 512, which is why long URLs from Deezer were
truncated when trying to report them to the SPI slave.

In addition, the definitions for registers 76, 79, and 239 specified a
maximum size of 1024, exceeding the internal buffer size. Writing such
long URLs by SPI slave is possible now, and attempting to do so
resulted in silent rejection because of the too small buffer size. Now
the register definitions match the buffer size of 2048.

Change-Id: I3ab9452c3eea51078a7abaef8bfa71c682200313