Convert TCP tunnel handling to C++.
Extension to more peers or multiple applink connections would have
become painful, so better switch to C++ to keep the level of pain low.
This new implementation can easily cope with many connections,
efficiently so.

Extension to queued TCP transactions is quite trivial now as well (we
will need this to get simultaneous writes from multiple clients
supported).

For practical purposes and to avoid DOS attacks over the network, we
restrict the number of TCP tunnels to be created by the ARM to a
maximum of 10, and allow up to 20 peer connections per tunnel. Note
that these restrictions are artificial and arbitrary, they are not
limitations imposed by the implementation itself.

Change-Id: I7deb670159476b8a4365d913c3d1c2611719b622