Fix network status register reported to SPI slave (WLAN).
The WPS and is-connecting states must be queried even in case a WLAN
connection is established already. This is because connecting via WPS
includes a WLAN scan in order to find a WPS registrar. While the scan
is in progress, we may still be connected to some WLAN network, so in
this situation we are connected and connecting at the same time.

Change-Id: I5a11040e19feec0e4e2dfa88a348ce6754c71f9e