Minor fix for start playing stream (#732).
We've sent de.tahifi.Streamplayer.Playback.Start() twice in case the
player hasn't started playing the first stream yet when we've received
the next stream URL from the ARM. The playback state used in this
context is the true playback state (not the pending state) as returned
by the de.tahifi.Streamplayer.URLFIFO.Push() method.

This is a "minor fix" because the stream player handles this situation
correctly: it knows it's going to play soon, so it won't try start
playing again. The fix doesn't really fix anything; still, dcpd
shouldn't have invoked the Start method the second time, hence the
fix.

Change-Id: I1e21aa091e9b70ac654d0a2be5ca57cf041f7b69