Dump DCP+DCPSYNC data sent to dcpspi to log.
For regular traffic over SPI, the log level must be set to "trace" to
see the logs. This is because dcpspi has its own (better) means of
dumping data to log, so the output of dcpd is probably only useful for
debugging communication between dcpd and dcpspi.

For traffic over TCP, the required log level is "normal".

Change-Id: If6e1840c0da155122f7843971b0dc585f45511ab