registers: New implementation for register 88.
The SPI slave may tell us now if the UPnP friendly name sent through
register 88 is the appliance's default name or if it is a name defined
by the user. The distinction is made on the last byte of the name: if
it is a binary 0, then the name is the appliance's default name.

This choice was made because current implementations of the ARM
software always send the friendly name without any zero terminator.
Our previous implementation didn't check for a zero terminator,
unfortunately, and writes a binary zero straight to our configuration
file. Therefore, we cannot just append an extra flag to the name.

By sheer luck, adding a binary zero to the end of the friendly name
doesn't confuse any of the UPnP clients tested, so we can append a
binary zero to make our case distinction while not messing up display
or sacrificing discoverability in case of the wrong mix of ARM and
Streaming Board.

The new implementation strips the trailing zero, of course.

The old behavior of this register is retained in case a protocol
version earlier than v1.0.6 is negotiated.

Change-Id: I43a573bc831463381dead311c029f827fff69082