registers: Always write FRIENDLY_NAME_GIVEN_BY_USER to config.
This is an improvement for register 88. In case we get an update via
protocol version 1.0.6, the value in FRIENDLY_NAME_GIVEN_BY_USER is a
meaningful one. Updating via the pre-v1.0.6 interface should also
update the FRIENDLY_NAME_GIVEN_BY_USER, even if only to a default
value, to keep the system behavior predictable.

Change-Id: Iddcd326fbd186365b1e6342b0a2d6766a8cdfe1f