Unit tests for inactive audio source, fixes.
We have a predefined audio source for inactive state now. This audio
source is preselected at early initialization time.

The reason for adding the previously implicit audio source as a
primary audio source is an inconsistency between first read-out of
register 81 and a later read-out after switching back and forth
between another source and inactive mode: on first read, zero bytes
were returned, on later reads a string on zero length was returned.
Now we always return the latter.

Change-Id: I263f54a8868fa007ff5d1bdabafa932975ab893d