Add options for audio sources.
There are no predefined options in dcpd, just a parser for the option
format. An option string can be appended to the name of the desired
audio source (also works for the empty, inactive audio source). The
option string is a comma-separated list of names and assignments.

Options can be either a simple flag named by the option name as it
appears in the option string, with its value set to TRUE, or they can
be an assignment so that the value is the string that follows an
equals sign. Escaping is done with backslash characters.

Change-Id: Ic92945c23fe13d535d0e5b47dc4bc08d0cf0722b