dcpregs_playstream: Use type-safe C++ stream IDs.
Also made some data members of the PlayAppStreamData and
PlayAnyStreamData structures private in an attempt to make the code a
bit more comprehensive.

Since there is still a legacy mechanism for detecting "app mode" in
place, I didn't go all the way and made all data private, but only for
those instances for which it was easy to do. The whole code probably
needs a good rewrite anyway...

Change-Id: If2936b0c4845bc3163c6bfb1aa5940bb1ddc834f