registers: Converted to C++.
Register definitions are proper objects now, and all register handlers
can be C++ functions now. This means that we are not forced to stick
with C-style header files anymore, and we can have namespaces.

A few other files had to be converted to C++ on the way, which is a
good thing.

Most code is still plain C compiled by a C++ compiler, but one of the
biggest C++ blockers has been removed now (the remaining big ones are
D-Bus and ConnMan code).

Change-Id: I6aa1b6c88a3d939d219f9f5282da92e3e1d986b5