Abandon BSD-style header inclusion order.
The BSD style guide prescribes a particular inclusion order of header
files (kernel headers, system headers first). This approach is widely
used, but today many projects take a different approach to improve the
quality of header files. We shall be no exception.

We now include local files first, with the .h/.hh file corresponding
to the .c/.cc file first. This adds a test for self-containedness of
header files (see also J. Lakos) and reduces the cost of long-term
maintainability.

Change-Id: Id5dd8b033d94620a98b34c67d28adeb31d3a08ff