configure: Detect doctest header file (#966).
New policy: Unit tests shall be written using the doctest framework.
Cutter is a royal PITA in many aspects; doctest is much easier to use
and still much more powerful.

We *must* get rid of Cutter and its dreaded Glib dependency to enable
ourselves to make sane use of Valgrind. We are currently tied to the
specific Glib version on srv-dev. This means that updating that
machine will break all (!) unit tests in all projects because of
outdated and unmaintainable Valgrind exceptions. This also means that
we cannot move the acceptance builds that are run for each commit to
our much faster srv-bob.

So, here is the first fragment that enables us to move forward.

Change-Id: I37a4c40519902bb7be47df6c13c9b1e7af7f1a52