connman: Function for parsing service names.
This is the first commit to make use of doctest. This is a *lot*
better.

To avoid interfering with the old Cutter tests, the new tests go to
tests_new. Once we have gathered a critical mass of new tests (by
adding new tests or by porting old tests to doctest), we'll rename
tests_new to tests and adapt the build job on Jenkins so that
Valgrind results, LCOV reports, and test statistics refer to the new
tests instead of the old ones.

Change-Id: Ic43faa63f1378bd9167ff8efbf05659cf8a3cf0e