ConnMan: Addresses can be checked and patched when set.
The Connman::Address<Connman::AddressType::MAC> is the first type to
make use of this: the setter checks if there are colons in the right
places, inserts colons if they are missing, and converts the passed
address string to uppercase. In case of any problems, an exception is
thrown so that the ctor works better now as well.

As a result, the process_address() function in network_device_list.cc
got significantly shorter and more comprehensible.

The other address types (IPv4, IPv6) do not have this smartness built
in yet, they still accept whatever is passed to them.

Change-Id: Ib10548d7ead27c12d9bbc591c3c5183d187d403c