Network: Make Connman::NetworkDeviceList::find() functions private.
Because they take std::string arguments, not typed MAC addresses. The
std::string is required because it is used as index into a std::map,
using Connman::Address<Connman::AddressType::MAC> would probably just
add boilerplate to Connman::Address to no avail.

The Connman::NetworkDeviceList::find() functions are only used
internally, so the impact of this change on the rest of the code is
zero.

Change-Id: I46f832cad7c1df92f49898aded25552761007cb9