Network prefs: Fix writing SSID and NetworkName keys.
In case any of the SSID or network name (but not both) are passed as
null pointers to network_prefs_put_wlan_config(), then this causes the
key corresponding to the string passed as null pointer to be deleted
from the configuration file now. Before, the corresponding key was
kept unaltered, leading to inconsistencies in case the configuration
request specified only one of the names.

Change-Id: I3c48eff68b23c90fa2b98ff064cc645bc1387100