Volume control: Fix bug found by cppcheck.
The public member function Mixer::VolumeControls::lookup() was never
implemented, only declared. Instead, the private version was picked up
by the linker, and it worked by straight coincidence. :-O

Cppcheck flagged a few warnings around this code starting with the
preceding commit, but it wasn't quite obvious what these were supposed
to mean. In the end it was bad overload resolution, and renaming the
internal lookup() function to lookup_rw() finally gave way to a linker
failure.

There are no kind words for this...

Change-Id: I770b367e59fc6f4c0d1678c53b8410b1c1b5a78d