Logged locks: Enable much better logs.
The application should use the LOGGED_LOCK_CONTEXT_HINT macro to leave
hints for the logged locks implementation. This will cause the
locations where the macro was invoked to be shown together with the
thread name in the log. The hints can be place anywhere, but they are
probably most useful right before a lock is about to be acquired.

Change-Id: I1ce5f897b44877c313bbb8a2c40e7c33c24efa26