File transfer: Fix compiler warning.
We used memset() on a non-trivial structure. Should have been safe,
but of course the compiler is always right in such a case. Ignoring
such a warning would be downright stupid.

The code became "a bit more" C++, but it is still mostly just C-style,
far from being nice and clean C++ code.

Change-Id: I4271539abbe716dda14d322b1ad41a21ecd35654