Player: Proper stop request handling.
In case the plain URL audio source is stopped via DRCP command, we
need to take care of stopping by ourselves and must not rely on drcpd
handling this situation. This is because we are in charge of
controlling the player in this case.

Stopping via drcpd caused us to see the stream player stop
notification as end of stream, so we skipped over to the next stream
just the way it should. Now we know that we have stopped it, so we can
react accordingly.

Change-Id: I34be65da2696927cf4d3634ae4789dd6e04590ec