Audio paths: Implement parser for AuPaL.
Audio Path Language... Our parser simply turns the specification into
an equivalent JSON object which is later going to be posted to the
audio path daemon.

We could also move the AuPaL parser to the audio path daemon and
process the input there. However, we turn these inputs into JSON
objects within DCPD already to get rid of DCP-specifics as soon as
possible. This allows us to deal with only a single audio path
specification format at system level, and avoids the question whether
or not AuPaL should also be understood by the REST API (*shudder*).

Change-Id: I6b9b35dba2658d7ca18be47a2cff5acc3003425d