transactions: Fix exception from hardened code.
We take the address of a data member in a vector, and we do it in a
way which is legal for plain arrays (take the address of the entry
beyond the end of the array). For vectors, this way isn't really legal
because operator[] is overloaded and may return something of which
the address cannot be taken, or whose address might not be as
unexpected.

Our fix is to use .data() to get at the plain underlying array of the
vector and do the address calculation on that array.

Change-Id: I3eed8c4cd7e21c9b383f8e51261d6f02c02f1aad