app playback: Emit stream info after successful push to player.
Stream information must not be sent around before we know for sure
that the stream has been placed into the stream player's queue.

Caring about "correct message ordering" on D-Bus from dcpd vs stream
player to end up in the UI in a meaningful order is wrong because we
cannot take any influence on how messages are scheduled on D-Bus
anyway (this might have been the initial reasoning behind the previous
implementation). Sending stream information for a stream which has not
made its way into the player, however, is definitely wrong. This
commit fixes this fault.

Change-Id: Ie2941db1922873b7a0cc270eff9ac34d4e236635