network config: Fix JSON serialization.
The WLAN security field is supposed to be an array of strings, not a
simple string. That's how it is documented in the REST API and how it
is delivered by ConnMan. Internally, we only store a single string for
historical reasons (because of our own configuration files), but this
is a detail that should not affect the REST API.

Another problem was a type error in hash computation which resulted in
an exception being thrown.

Change-Id: I86cfc601b4b5441582bdcd182945b4f2766c626d