dcp: Protocol version is 1.1.0 now.
This is because we need to distinguish update register behaviors
between V1 and V2 series. The V2 series start with 1.1.0 so that the
SPI slave has an easier way to figure out how it should talk with the
Streaming Board.

If new version of this software needs to be used on V1 boards, then we
must branch and set the protocol version back to 1.0.x.

Change-Id: I025fb5be525541794b381eab9dfbdc1933b84985