audiosources: Avoid D-Bus deadlock on audio source selection (#1455).
When TAPSwitch tells us that an audio source is available, we ask
TAPSwitch about the source details. This is done by making a
synchronous D-Bus call, and it is made directly from D-Bus context.

This works, unless we are told about our own source "strbo.plainurl"
and we are in the middle of an audio source request for that source.
In this situation, the synchronous D-Bus call leads to a deadlock, and
thus a D-Bus timeout. The system still moves on and plays the given
URL, but there is an unnecessarily long delay before it does.

This commit moves the synchronous D-Bus call to the main loop context,
thus avoiding the deadlock.

Change-Id: I28b6085c3836bc02dcdf4c184631a12dd7fccf70