dcp: Return empty media services list on D-Bus timeout (#1440).
This patch is for register 106.

A D-Bus timeout can occur in case the Airable list broker is busy
waiting for a TCP/IP timeout (a few minutes, maybe). We let the SPI
slave know that there are no information (yet) in this case instead of
having it run into a timeout of its own.

Change-Id: I3d0c450d99f8d20103129444204a653576bef3e6